new(xychart): move all Animated exports to @visx/xychart/animated
