+++ packages/visx-xychart/src/animated/index.ts
@@ -1,0 +1,42 @@
+// components
+export { default as Annotation } from '../components/annotation/AnimatedAnnotation';
+export { default as AnnotationLabel } from '../components/annotation/AnnotationLabel';
+export { default as AnnotationConnector } from '../components/annotation/AnnotationConnector';
+export { default as AnnotationCircleSubject } from '../components/annotation/AnnotationCircleSubject';
+export { default as AnnotationLineSubject } from '../components/annotation/AnnotationLineSubject';
+export { default as Axis } from '../components/axis/AnimatedAxis';
+export { default as Grid } from '../components/grid/AnimatedGrid';
+export { default as Tooltip } from '../components/Tooltip';
+export { default as XYChart } from '../components/XYChart';
+
+// series components
+export { default as AreaSeries } from '../components/series/AnimatedAreaSeries';
+export { default as BarSeries } from '../components/series/AnimatedBarSeries';
+export { default as BarStack } from '../components/series/AnimatedBarStack';
+export { default as BarGroup } from '../components/series/AnimatedBarGroup';
+export { default as GlyphSeries } from '../components/series/AnimatedGlyphSeries';
+export { default as LineSeries } from '../components/series/AnimatedLineSeries';
+
+// context
+export { default as DataContext } from '../context/DataContext';
+export { default as EventEmitterContext } from '../context/EventEmitterContext';
+export { default as ThemeContext } from '../context/ThemeContext';
+export { default as TooltipContext } from '../context/TooltipContext';
+
+// providers
+export { default as DataProvider } from '../providers/DataProvider';
+export { default as EventEmitterProvider } from '../providers/EventEmitterProvider';
+export { default as ThemeProvider } from '../providers/ThemeProvider';
+export { default as TooltipProvider } from '../providers/TooltipProvider';
+
+// hooks
+export { default as useEventEmitter } from '../hooks/useEventEmitter';
+
+// themes
+export { default as lightTheme } from '../theme/themes/light';
+export { default as darkTheme } from '../theme/themes/dark';
+export { default as buildChartTheme } from '../theme/buildChartTheme';
+export { allColors, grayColors, defaultColors } from '../theme/colors';
+
+// types
+export * from '../types';
