+++ packages/visx-xychart/test/components/AreaSeries.test.tsx
@@ -2,7 +2,8 @@ import React, { useContext, useEffect } from 'react';
 import { animated } from 'react-spring';
 import { mount } from 'enzyme';
 import { Area, LinePath } from '@visx/shape';
-import { AnimatedAreaSeries, DataContext, AreaSeries, useEventEmitter } from '../../src';
+import { DataContext, AreaSeries, useEventEmitter } from '../../src';
+import { AreaSeries as AnimatedAreaSeries } from '../../src/animated';
 import getDataContext from '../mocks/getDataContext';
 import setupTooltipTest from '../mocks/setupTooltipTest';
 import { XYCHART_EVENT_SOURCE } from '../../src/constants';
