+++ packages/visx-xychart/test/components/LineSeries.test.tsx
@@ -2,7 +2,8 @@ import React, { useContext, useEffect } from 'react';
 import { animated } from 'react-spring';
 import { mount } from 'enzyme';
 import { LinePath } from '@visx/shape';
-import { AnimatedLineSeries, DataContext, LineSeries, useEventEmitter } from '../../src';
+import { DataContext, LineSeries, useEventEmitter } from '../../src';
+import { LineSeries as AnimatedLineSeries } from '../../src/animated';
 import getDataContext from '../mocks/getDataContext';
 import setupTooltipTest from '../mocks/setupTooltipTest';
 import { XYCHART_EVENT_SOURCE } from '../../src/constants';
